Created a proposal notebook for the group project
Started to write the introduction, covered all the needed topics required in the intro. Made a skeleton outline of the proposal and all the things that need to be included in each sections. # things out that we need to do and wrote that we might need to consider talking about in the proposal.